Moved GUI code to separate class. ClientMain calls GUI launch method in GUI class, which holds all the GUI code now.
